Refactor actions, separating out types and utils
